Implement syncing of playback when pressing play
As a listener, you can now sync your playback when a broadcaster is
active (either playing or paused) by pressing the play button.

TODO: The listener's web player will play even if the broadcaster is
paused, the listener should not play but insead seek to the same time as
the broadcaster and pause. Spotify does not provide a direct API for
doing so.